# src/main/resources/api/openapi.yml
openapi: '3.0.3'
info:
  title: Book Club Reading Tracker API
  description: >
    API for managing an online book club's reading progress, schedules, and discussion sessions.
    It allows the creation of books, reading schedules, members, member progress, and meeting sessions.
  version: '1.0'
servers:
  - url: http://localhost:8080/v1
    description: Local server
  - url: https://api.example.com/v1
    description: Production server
...
paths:
...
  /members:
    get:
      tags:
        - Members
      summary: Return a list of members with pagination
      operationId: listMembers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - in: query
          name: name
          description: Filter members by (case-insensitive) name (optional)
          required: false
          schema:
            type: string
            example: "John"
        - in: query
          name: email
          description: Filter members by email (optional)
          required: false
          schema:
            type: string
            example: "john@example.com"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMembers'
    post:
      tags:
        - Members
      summary: Create a new member
      operationId: createMember
      requestBody:
        description: Member to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Bad request
  /members/{memberId}:
    get:
      tags:
        - Members
      summary: Get a member by their ID
      operationId: getMemberById
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: Member not found
    put:
      tags:
        - Members
      summary: Update a member by their ID
      operationId: updateMember
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        description: Updated member information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: Bad request
        '404':
          description: Member not found
    delete:
      tags:
        - Members
      summary: Delete a member by their ID
      operationId: deleteMember
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '204':
          description: No Content
        '404':
          description: Member not found
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number (default is 0)
      required: false
      schema:
        type: integer
        default: 0
        format: int32
        example: 0
    PerPage:
      name: perPage
      in: query
      description: NNumber of items per page (default is 10)
      required: false
      schema:
        type: integer
        default: 10
        format: int32
        example: 10
...
    MemberId:
      name: memberId
      in: path
      description: ID of the member to retrieve
      required: true
      schema:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440000"
...
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "John Doe"
          minLength: 3
          maxLength: 100
        email:
          type: string
          example: "john.doe@example.com"
          maxLength: 255
          format: email
...
    PaginatedMembers:
      type: object
      properties:
        page:
          type: integer
          example: 0
          format: int32
        perPage:
          type: integer
          example: 10
          format: int32
        totalPages:
          type: integer
          example: 5
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
    CreateMemberRequest:
      type: object
      properties:
        name:
          type: string
          example: "John Doe"
          minLength: 3
          maxLength: 100
        email:
          type: string
          example: "john.doe@example.com"
          format: email
          maxLength: 255
      required:
        - name
        - email
    UpdateMemberRequest:
      type: object
      properties:
        name:
          type: string
          example: "John Doe"
          minLength: 3
          maxLength: 100
        email:
          type: string
          example: "john.doe@example.com"
          maxLength: 255
          format: email
      required:
        - name
        - email
...